test: create folder for e2e, add unit test for personality
