+++ fortunetelling/src/__test__/e2e/README.md
