+++ fortunetelling/src/__test__/e2e/samplejest.test.js
