+++ fortunetelling/src/__test__/e2e/samplemobiletest.js
