--- fortunetelling/src/__test__/sampleunit.test.js
@@ -1,18 +1,0 @@
-// const {getHoroscope} = require('fortunetelling/src/js/genHoroscope.js');
-/*
-test('getHoroscope', () => {
-
-    const message1 = getHoroscope(1); // Use any sign number for testing
-
-
-    const message2 = getHoroscope(5);
-
-    // Expect the messages to be different
-    expect(message1).not.toBe(message2);
-
-    // In order to test whether two consecutive days produce the same message, Aaron needs to change the parameter taken in
-});
-*/
-test('Sample', () => {
-	expect(1 + 2).toBe(3);
-});
