+++ fortunetelling/src/__test__/unit/testSpeechify.js
@@ -1,0 +1,8 @@
+/*
+import { ok } from 'node:assert';
+import { it } from './helper.js';
+
+// Import functions to test
+import { Speechify } from '../../js/speechify.js';
+import { getJSON } from '../../js/utils.js';
+*/
